requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dj-database-url==2.3.0
Django==5.1.6
django-allauth==65.3.1
django-environ==0.12.0
djangorestframework==3.15.2
docutils==0.21.2
environ==1.0
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
htmx==0.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
markdown2==2.5.3
MarkupSafe==3.0.2
multidict==6.1.0
nicegui==2.11.1
orjson==3.10.15
pillow==11.1.0
propcache==0.3.0
pscript==0.7.7
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==1.10.21
Pygments==2.19.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vbuild==0.8.2
watchfiles==1.0.4
websockets==15.0
wsproto==1.2.0
yarl==1.18.3
